requirements.txt
opencv-python==4.7.0.68
mediapipe==0.10.5
scikit-learn==1.6.1
Pillow==8.4.0
arabic-reshaper==2.1.3
python-bidi==0.4.2
numpy=2.0.2
